+++ src/iotvm-eventengine/src/main/java/org/softwareforce/iotvm/eventengine/cep/ct/AverageCalculationCompositeTransformationParameters.java
@@ -171,6 +171,7 @@ public class AverageCalculationCompositeTransformationParameters
   }
 
   // TODO Temporarily enabled.
+  // TODO Not sure!
   public int getFutureWindowsLookup() {
     return futureWindowsLookup;
   }